query corresponding fs file/chunks with right fs file id

+++ gridfs_file_chunk.py
@@ -26,22 +26,37 @@ conn = connect('test', host='localhost', port=27017)
 # 使用mongoengine GridFS存文件
 filename = 'a pic'
 report = ReportFile(filename=filename, datetime=datetime.datetime.now())
-fh = open('qinghai_lake.jpg', 'rb')
-report.report_file.put(fh, content_type='img/jpeg')
+with open('qinghai_lake.jpg', 'rb') as fh:
+    report.report_file.put(fh, content_type='img/jpeg')
 report.save()
 
 # 读取文件
 report = ReportFile.objects(filename=filename).first()
+if not report:
+    print 'Report file not found in the DB!'
+    exit(1)
+# 对应的fs file id
+fid = report.report_file._id
 report_file = report.report_file.read()
 with open('retr_img.jpg', 'w') as retr_fh:
     retr_fh.write(report_file)
 conn.disconnect()
 
+# 使用pymongo来查询对应的fs files和fs chunks
 conn = Connection('localhost', 27017)
 db = conn.test
-for each_file in db.fs.files.find():
-    print each_file
-print db.fs.chunks.find().count()
+fs_file = db.fs.files.find_one({'_id': fid})
+if not fs_file:
+    print 'The fs file does not exist in collection fs.files!'
+    exit(1)
+print 'The fs file id is \'%s\'' % fs_file['_id']
+chunk_count = db.fs.chunks.find({'files_id': fid}).count()
+if chunk_count == 0:
+    print 'There is no any fs chunk for the stored large file!'
+    exit(1)
+print 'There are %d fs chunks for the stored large file.' % chunk_count
+
+# drop collection
 db.drop_collection('TestFile')
 db.drop_collection('fs.files')
 db.drop_collection('fs.chunks')
